+++ CHANGELOG.md
@@ -22,10 +22,10 @@ Changes since v2.38.1
 ### Fixes
 - Resolved issue where workflow voting could not be removed, which caused UI to display raw translation keys due to nil voting values. (#3357)
 - Fixed an issue with the application page where the 'Add member' dropdown menu displayed members that already exist in the application.  (#1412)
-- Document title no longer shows an extra em dash when the localized application title is empty. 
 - Document title no longer shows an extra em dash when the localized application title is empty. (#3398)
+- Fix audit log API. Technical error from incorrect response. (#3380)
 - Fixed an issue causing the `fix-userid` administration script to break in some edge cases (#3383)
- 
+
 ## v2.38.1 "Välimerenkatu +1" 2025-01-27
 
 ### Fixes